conn/usb: Shadowed error can lead to null pointer deref
When the access to the usb device failes given an explicit usb device it can lead to the error variable
err being shadowed. This is now followed by a `goto handleError` which will return the original (nil)
err instead of the thrown one.
```
	ser, err = ptouchgo.Open("usb://0x2061", 0, true)
```
Here err can never be non nil. We therefore must not shadow the original error for the codeflow to work.